findings_extractor/requirements.txt
pypdf
pandas
openpyxl
deepseek
ollama